Refactored SvnOperationDialog to be just a container which can run multiple SvnOperationProcesses, and converted all existing ops to this new model

+++ dualitysvn/SvnOperationProcess.py
@@ -187,6 +187,10 @@ class SvnOperationProcess:
 		
 	# Abort process prematurely
 	def endProcess(self):
+		# sanity check: only need to do this for "running" processes
+		if self.status != ProcessStatus.STATUS_WORKING:
+			return;
+		
 		# kill process - only way to get rid of console apps on windows
 		self.process.kill();
